+++ src/modules/freshrelevance/store.ts
@@ -1,0 +1,6 @@
+import { Module } from 'vuex';
+
+export const freshRelevanceStore: Module<{}, any> = {
+  namespaced: true,
+  state: {}
+};