+++ src/modules/freshrelevance/types.ts
@@ -1,0 +1,4 @@
+import * as Vuex from 'vuex';
+import RootState from '@vue-storefront/core/types/RootState';
+
+export type Store = Vuex.Store<RootState>;